requirements.txt
flask==2.2.5
Flask-Session==0.4.0
python-dotenv==1.0.0

pdfplumber==0.8.1
python-docx==0.8.11
python-pptx==0.6.21

spacy==3.7.1
sentence-transformers==2.2.2
keybert==0.6.0
transformers==4.37.0
torch==2.2.2

faiss-cpu==1.7.4.post2  # optional for faster similarity (if pip supports your platform)
sentencepiece==0.1.99

reportlab==4.0.0
